backend/requirements.txt
Django==4.2.7
strawberry_graphql_django==0.25.0
django_cors_headers==4.3.1
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
pandas==2.1.3
sounddevice==0.4.6
openpyxl==3.1.2
omegaconf==2.3.0
soundfile==0.12.1